# vindexes: Efficient unicode hashing