--- a/go/mysql/collations/vindex/collate/golden_test.go
+++ b/go/mysql/collations/vindex/collate/golden_test.go
@@ -1,7 +1,0 @@
-package collate
-
-import "testing"
-
-func TestGenerateGolden(t *testing.T) {
-
-}