--- a/go/mysql/collations/vindex/doc.go
+++ b/go/mysql/collations/vindex/doc.go
@@ -1,0 +1,16 @@
+// Copyright 2015 The Go Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file.
+
+//go:generate go run gen.go
+
+// text is a repository of text-related packages related to internationalization
+// (i18n) and localization (l10n), such as character encodings, text
+// transformations, and locale-specific text handling.
+//
+// There is a 30 minute video, recorded on 2017-11-30, on the "State of
+// vitess.io/vitess/go/mysql/collations/vindex" at https://www.youtube.com/watch?v=uYrDrMEGu58
+package text
+
+// TODO: more documentation on general concepts, such as Transformers, use
+// of normalization, etc.