--- a/go/mysql/collations/vindex/unicode/doc.go
+++ b/go/mysql/collations/vindex/unicode/doc.go
@@ -1,0 +1,8 @@
+// Copyright 2015 The Go Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file.
+
+// unicode holds packages with implementations of Unicode standards that are
+// mostly used as building blocks for other packages in vitess.io/vitess/go/mysql/collations/vindex,
+// layout engines, or are otherwise more low-level in nature.
+package unicode