--- a/go/unicode2/unicode.go
+++ b/go/unicode2/unicode.go
@@ -1,3 +1,7 @@
+// Copyright 2011 The Go Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file.
+
 package unicode2
 
 import "unicode"