… go/mysql/collations/colldata/golden_test.go
… go/mysql/collations/integration/collations_test.go
… go/mysql/collations/integration/weight_string_test.go
… go/mysql/collations/testutil/golden.go
… go/mysql/collations/testutil/lang.go
… go/mysql/collations/vindex/.gitignore
… go/mysql/collations/vindex/internal/colltab/trie.go
… go/mysql/collations/vindex/unicode/norm/composition.go
… go/mysql/collations/vindex/unicode/norm/forminfo.go
… go/mysql/collations/vindex/unicode/norm/input.go
… go/mysql/collations/vindex/unicode/norm/iter.go
… go/mysql/collations/vindex/unicode/norm/normalize.go
… go/mysql/collations/vindex/unicode/norm/tables15.0.0.go
… go/mysql/collations/vindex/unicode/norm/trie.go
… go/mysql/json/json_path.go
… go/textutil/strings.go
… go/tools/asthelpergen/asthelpergen.go
… go/vt/mysqlctl/backup.go
… go/vt/vtgate/vindexes/hash_test.go
… go/vt/vtgate/vindexes/lookup_unicodeloosemd5_hash.go
… go/vt/vtgate/vindexes/testdata/unicode_hash_golden.json
… go/vt/vtgate/vindexes/unicode.go
… go/vt/vtgate/vindexes/unicode_hash_test.go
… go/vt/vtgate/vindexes/unicode_test.go
… go/vt/vtgate/vindexes/unicodeloosemd5.go
… go/vt/vtgate/vindexes/unicodeloosexxhash.go