@dbussink 2023-10-30
> This can be done much more efficiently because the value we require for a vindex is always a hash, so it can be calculated using a constant amount of memory if instead of generating a weight string and hashing it, we write directly into a hasher.

The vendoring also addresses another issue. When `x/text` is updated, the collation tables can be fixed for a newer Unicode version but that could (very subtlety) break a vindex since it changes what input hashes to.

Here we vendor, so we essentially set it in stone and keep it always consistent for the future.