@vmg 2023-10-31
Yes, this is 100% backwards compatible and there are golden tests that verify it. Otherwise we'd break all existing Vitess deployments by changing the shard position of existing values!